+++ src/sys_deno/ports/deps.ts
@@ -1,5 +1,1 @@
-//! dependencies used by the cli
-
 export * from "../../deps.ts";
-
-export { Table } from "https://deno.land/x/cliffy@v1.0.0-rc.4/table/table.ts";
